+++ docs/changelog.md
@@ -3,6 +3,7 @@
 * Added Prysm CL (beacon, validator) Launcher
 * Made EL & CL client log levels configurable as a module param, `logLevel`
 * Added self-documenting code for module params
+* When an invalid EL or CL client type is provided in the params, the valid values are printed to the user
 
 ### Changes
 * The `WaitForBeaconClientAvailability` method also checks if the returned status is READY, which means the node is synced